Add: Changed resolution multipliers (closes #54) (see more)
New resolution multipliers: x2, x2.5, x3, x3.5, x4, x5, x6, x8
Multipliers use the base 256x224 resolution to work
Also checks for your monitor size (sf::VideoMode::getDesktopMode().height)
to check if you're able to fit the resolution (for example, a 1080p screen
won't be able to use x8 multiplier)

Resolutions are: 512x448, 640x560, 768x672, 896x784, 1024x896, 1280x1120, 1536x1344 and 2048x1792, respectively

+++ src/game.cpp
@@ -125,7 +125,7 @@ void Game::updateResolution() {
     if (window.isOpen()) {
         window.close();
     }
-    unsigned int resolutionMultiplier = Settings::getResolutionMultiplier();
+    float resolutionMultiplier = Settings::getResolutionMultiplier();
     window.create(sf::VideoMode(baseWidth * resolutionMultiplier,
                                 baseHeight * resolutionMultiplier),
                   "Super Mario Kart", WINDOW_STYLE);
